// src/app/lib/types.ts
export interface selectType {
  id: string | number;
  value: string;
}
// ...
export interface CostCenterType {
  idCostCenter?: number;
  uniqueId?: string;
  costCenter: string;
  description: string;
  bidang: string;
  dinas: string;
  directorat: string;
  groupDinas: string;
  profitCenter: string;
  active: boolean;
  createdAt?: Date;
  createdBy?: string;
  updatedAt?: Date;
  updatedBy?: string | null;
}

export interface UploadRKAP {
  years: string;
  file: File;
  createdBy: string;
}
export interface personalSummaryTypes {
  totalSubmissionValue: number;
  data: personalSummaryDataTypes[];
}
export interface personalSummaryDataTypes {
  idRealization: string;
  dinas: string;
  month: number;
  years: number;
  requestNumber: string | null;
  typeSubmission: string;
  submissionValue: number;
  status: string;
  requestBy: string;
  responsibleOfRequest: string;
  description: string;
}